requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psd-tools>=1.9.0
werkzeug>=3.0.0
